Llama Index/requirements.txt
llama-index>=0.10.0
llama-index-llms-huggingface-api>=0.1.0
llama-index-embeddings-huggingface>=0.1.0
llama-index-vector-stores-chroma>=0.1.0
chromadb>=0.4.0
python-dotenv>=1.0.0
